Fix: Restore UI stability and Camera functionality after rollback
- CameraBehaviour: Resolved syntax error in HandleZoom caused by merge conflict.
- CameraBehaviour: Updated IsMouseOverUI to correctly block clicks when interacting with topStatsHUD and taskPanel.
- CharacterStats: Refactored UpdateHealth to use the Health property, ensuring stat changes correctly trigger UI update events.
- UIManager: Restored critical null safety checks in UpdateCharacterStatsDisplay and UpdateTimeDisplay to prevent UI logic crashes from unassigned fields.
- InventoryItemUI: Added null checks to Setup method for safer prefab instantiation.

+++ Assets/Scripts/SaveStructure/SaveDataClasses.cs
@@ -20,7 +20,7 @@ public class CharacterSaveData
 {
     public string name;
     public string iconName;
-    public float health;
+    public int health;
     public int stability;
     public int learning;
     public int workReadiness;
